# app/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Next.js API Documentation",
    "description": "API documentation for Next.js application",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
    "/hello": {
      "get": {
        "summary": "Hello World endpoint",
        "description": "Returns a greeting message",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Hello from the API!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/animals/random": {
      "get": {
        "summary": "Get random animals",
        "description": "Returns one or more random animals",
        "parameters": [
          {
            "name": "count",
            "in": "query",
            "description": "Number of random animals to return (1-100)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "oneOf": [
                        {
                          "type": "string",
                          "example": "Lion",
                          "description": "Single random animal when no count specified"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": ["Lion", "Tiger", "Bear"],
                          "description": "Array of random animals when count specified"
                        }
                      ]
                    },
                    "total": {
                      "type": "integer",
                      "example": 1,
                      "description": "Number of animals returned"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid query parameters"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}